% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecg_filter.R
\name{ecg_filter}
\alias{ecg_filter}
\title{ecg_filter}
\usage{
ecg_filter(sample, sample_rate = 500, butter_low = 5, butter_high = 80)
}
\arguments{
\item{sample}{numeric ecg register}

\item{sample_rate}{acquisition rate in Hz}

\item{butter_low}{low-pass butterfilter param Hz}

\item{butter_high}{high-pass butterfilter param Hz}
}
\value{
Numeric vector with filtered data
}
\description{
Filter ECG noise and establish a baseline with butterfilter
}
\examples{
data <- ECGexample1
data$filtered <- ecg_filter(data$raw_signal,sample_rate = 500, butter_low = 5, butter_high = 80)

}
