% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECGexample1.R
\docType{data}
\name{ECGexample1}
\alias{ECGexample1}
\title{20 seconds ECG raw data (acquisition rate 500Hz) obtained from PhysioNet database}
\format{
A data frame with 10000 rows and 2 variables:
\describe{
\item{index}{numeric index}
\item{raw_signal}{ECG data vector (mV)}
...
}
}
\source{
\url{https://physionet.org/content/ecgiddb/1.0.0/}
}
\usage{
ECGexample1
}
\description{
20 seconds ECG raw data (acquisition rate 500Hz) obtained from PhysioNet database
}
\keyword{datasets}
