% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ecg_features.R
\name{extract_ecg_features}
\alias{extract_ecg_features}
\title{Extract ECG features from a processed ECG data frame}
\usage{
extract_ecg_features(ecg_df)
}
\arguments{
\item{ecg_df}{A data frame returned by read_ecg() or read_ecg_csv().}
}
\value{
A structured list of ECG features.
}
\description{
Extract ECG features from a processed ECG data frame
}
