% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_r_peaks.R
\name{find_r_peaks}
\alias{find_r_peaks}
\title{find_r_peaks}
\usage{
find_r_peaks(sample, FCmax = 220, sample_rate = 500, threshold = 0.2)
}
\arguments{
\item{sample}{numeric ecg register}

\item{sample_rate}{acquisition rate in Hz}

\item{threshold}{set minimum peak value, can be adjusted to avoid T-waves identification}
}
\value{
Numeric vector with peak indexes
}
\description{
Find R peaks from a ECG sample, strongly recommended to use data obtained from ecg_filter(),
can be checked with plot_r_peaks()
}
\examples{
data <- ECGexample1
data$filtered <- ecg_filter(data$raw_signal,sample_rate = 500, butter_low = 5, butter_high = 80)
peaks <- find_r_peaks(data$filtered, FCmax = 220, sample_rate = 500, threshold = .2)

}
