% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrv_time_analysis.R
\name{hrv_time_analysis}
\alias{hrv_time_analysis}
\title{hrv_time_analysis}
\usage{
hrv_time_analysis(sample, Rpeaks, sample_rate = 500)
}
\arguments{
\item{sample}{numeric ecg register}

\item{Rpeaks}{numeric vector with R peak indexes (can be obtained from find_r_peaks())}

\item{sample_rate}{acquisition rate in Hz}
}
\value{
List with the quantified parameters for the analysis, could have missing information if the sample is small
}
\description{
Heart Rate Variability (HRV) time domain analysis from the sample (SDNN, pnn50, rMSSD, ...)
}
\examples{
data <- ECGexample1
data$filtered <- ecg_filter(data$raw_signal,sample_rate = 500, butter_low = 5, butter_high = 80)
peaks <- find_r_peaks(data$filtered, FCmax = 220, sample_rate = 500, threshold = .2)
hrv_time_analysis(data$filtered, peaks)

}
