% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_ecg_features.R
\name{interpret_ecg_features}
\alias{interpret_ecg_features}
\title{Interpret extracted ECG features}
\usage{
interpret_ecg_features(features)
}
\arguments{
\item{features}{A list returned by extract_ecg_features()}
}
\value{
A structured interpretation of the ECG features
}
\description{
Interpret extracted ECG features
}
