% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_hrv.R
\name{interpret_hrv}
\alias{interpret_hrv}
\title{Generate a plain-English interpretation of HRV results}
\usage{
interpret_hrv(hrv)
}
\arguments{
\item{hrv}{An \code{ecg_hrv} object from \code{compute_hrv()}.}
}
\value{
A named list with interpretation text sections.
}
\description{
Generate a plain-English interpretation of HRV results
}
