% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_heart_rate.R
\name{mean_heart_rate}
\alias{mean_heart_rate}
\title{mean_heart_rate}
\usage{
mean_heart_rate(sample, Rpeaks, sample_rate = 500)
}
\arguments{
\item{sample}{numeric ecg register}

\item{Rpeaks}{numeric vector with R peak indexes (can be obtained from find_r_peaks())}

\item{sample_rate}{acquisition rate in Hz}
}
\value{
Float with the mean heart rate from the sample
}
\description{
Identify the mean heart rate from an ECG sample
}
\examples{
data <- ECGexample1
data$filtered <- ecg_filter(data$raw_signal,sample_rate = 500, butter_low = 5, butter_high = 80)
peaks <- find_r_peaks(data$filtered, FCmax = 220, sample_rate = 500, threshold = .2)
mean_heart_rate(data$filtered, peaks)

}
