% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecg.R
\name{plot_ecg}
\alias{plot_ecg}
\title{Plot an ECG waveform with annotated beats}
\usage{
plot_ecg(ecg_df, lead = "MLII", show_annotations = TRUE)
}
\arguments{
\item{ecg_df}{An \code{ecg_df} object from \code{read_ecg()}.}

\item{lead}{Which lead to plot. Default "MLII".}

\item{show_annotations}{If TRUE, mark beat positions with colored points.}
}
\value{
A ggplot2 object.
}
\description{
Plot an ECG waveform with annotated beats
}
\examples{
\dontrun{
df <- read_ecg("data/100", from_sec = 0, to_sec = 10)
plot_ecg(df)
plot_ecg(df, lead = "V5")
}

}
