% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_r_peaks.R
\name{plot_r_peaks}
\alias{plot_r_peaks}
\title{plot_r_peaks}
\usage{
plot_r_peaks(sample, Rpeaks)
}
\arguments{
\item{sample}{numeric ecg register}

\item{Rpeaks}{numeric vector with R peak indexes (can be obtained from find_r_peaks())}
}
\value{
Plot from the sample with R peak identification
}
\description{
Simple plot for visual identification of the R peaks that can be obtained from find_r_peaks()
Helpfull for visualization and parameters settings for the functions ecg_filter() and find_r_peaks()
}
\examples{
data <- ECGexample1
data$filtered <- ecg_filter(data$raw_signal,sample_rate = 500, butter_low = 5, butter_high = 80)
peaks <- find_r_peaks(data$filtered, FCmax = 220, sample_rate = 500, threshold = .2)
plot_r_peaks(data$filtered, peaks)

}
