% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rr.R
\name{plot_rr}
\alias{plot_rr}
\title{Plot RR intervals over time}
\usage{
plot_rr(hrv, highlight_threshold = 150)
}
\arguments{
\item{hrv}{An \code{ecg_hrv} object from \code{compute_hrv()}.}

\item{highlight_threshold}{Numeric. Flag RR intervals deviating more
than this many ms from the median. Default 150.}
}
\value{
A ggplot2 object.
}
\description{
Plot RR intervals over time
}
