% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ecg.R
\name{print.ecg_df}
\alias{print.ecg_df}
\title{Print basic information about an ECG data frame}
\usage{
\method{print}{ecg_df}(x, ...)
}
\arguments{
\item{x}{An \code{ecg_df} object returned by \code{read_ecg()}.}

\item{...}{Further arguments passed to \code{print.data.frame}.}
}
\description{
Print basic information about an ECG data frame
}
