% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_hrv.R
\name{print.ecg_hrv}
\alias{print.ecg_hrv}
\title{Print a summary of HRV results}
\usage{
\method{print}{ecg_hrv}(x, ...)
}
\arguments{
\item{x}{An \code{ecg_hrv} object from \code{compute_hrv()}.}

\item{...}{Ignored.}
}
\description{
Print a summary of HRV results
}
